pyproject.toml
[tool.poetry]
name = "shiny-octo-spork"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
trio = "^0.19.0"
asks = "^2.4.12"
python-dotenv = "^0.18.0"
redio = "^0.5.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "^0.910"
wemake-python-styleguide = "^0.15.3"
pytest-coverage = "^0.0"
pytest-trio = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
